reworked cmake file split database plugins files
