+++ osquery/database/CMakeLists.txt
@@ -6,35 +6,44 @@
 #  in the COPYING file in the root directory of this source tree).
 #  You may select, at your option, one of the above-listed licenses.
 
-ADD_OSQUERY_LIBRARY(TRUE osquery_database
-  database.cpp
-  plugins/ephemeral.cpp
+ADD_OSQUERY_LIBRARY(TRUE osquery_database "")
+
+target_sources(osquery_database
+  PRIVATE
+    "${CMAKE_CURRENT_LIST_DIR}/database.cpp"
+    "${CMAKE_CURRENT_LIST_DIR}/plugins/ephemeral.cpp"
+    "${CMAKE_CURRENT_LIST_DIR}/plugins/ephemeral.h"
 )
 
-# Optionally (but by default), add the RocksDB database plugin.
-add_definitions(-DROCKSDB_LITE=1)
-
-# Begin with just the SQLite database plugin.
-set(OSQUERY_DATABASE_PLUGINS
-  plugins/sqlite.cpp
-  plugins/rocksdb.cpp
+ADD_OSQUERY_TEST(TRUE
+  "${CMAKE_CURRENT_LIST_DIR}/tests/database_tests.cpp"
+  "${CMAKE_CURRENT_LIST_DIR}/tests/plugin_tests.cpp"
+  "${CMAKE_CURRENT_LIST_DIR}/tests/plugin_tests.h"
+  "${CMAKE_CURRENT_LIST_DIR}/tests/results_tests.cpp"
 )
 
-set(OSQUERY_DATABASE_PLUGIN_TESTS
-  database/tests/plugin_tests.cpp
-  database/plugins/tests/sqlite_tests.cpp
-  database/plugins/tests/rocksdb_tests.cpp
+ADD_OSQUERY_LIBRARY(FALSE osquery_database_plugins "")
+
+target_sources(osquery_database_plugins
+  PRIVATE
+    "${CMAKE_CURRENT_LIST_DIR}/plugins/rocksdb.cpp"
+    "${CMAKE_CURRENT_LIST_DIR}/plugins/rocksdb.h"
+    "${CMAKE_CURRENT_LIST_DIR}/plugins/sqlite.h"
+    "${CMAKE_CURRENT_LIST_DIR}/plugins/sqlite.cpp"
 )
 
-# Plugins (which do not include the shim/ephemeral) are 'additional'.
-ADD_OSQUERY_LIBRARY(FALSE osquery_database_plugins ${OSQUERY_DATABASE_PLUGINS})
+target_compile_definitions(osquery_database_plugins
+  PRIVATE
+    "-DROCKSDB_LITE=1"
+)
 
-# Non-plugin tests are core.
-file(GLOB OSQUERY_DATABASE_TESTS "tests/*.cpp")
-ADD_OSQUERY_TEST(TRUE ${OSQUERY_DATABASE_TESTS})
+ADD_OSQUERY_TEST(FALSE
+  "${CMAKE_CURRENT_LIST_DIR}/tests/plugin_tests.cpp"
+  "${CMAKE_CURRENT_LIST_DIR}/tests/plugin_tests.h"
+  "${CMAKE_CURRENT_LIST_DIR}/plugins/tests/rocksdb_tests.cpp"
+  "${CMAKE_CURRENT_LIST_DIR}/plugins/tests/sqlite_tests.cpp"
+)
 
-# Plugin tests are additional.
-ADD_OSQUERY_TEST(FALSE ${OSQUERY_DATABASE_PLUGIN_TESTS})
-
-file(GLOB OSQUERY_DATABASE_BENCHMARKS "benchmarks/*.cpp")
-ADD_OSQUERY_BENCHMARK(${OSQUERY_DATABASE_BENCHMARKS})
+ADD_OSQUERY_BENCHMARK(
+  "${CMAKE_CURRENT_LIST_DIR}/benchmarks/database_benchmarks.cpp"
+)
